Move Attachment to own package
