+++ src/main/java/seedu/address/model/person/Person.java
@@ -10,6 +10,7 @@ import java.util.Objects;
 import java.util.Set;
 
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.model.attachment.Attachment;
 import seedu.address.model.tag.Tag;
 
 /**
