+++ src/test/java/seedu/address/model/person/AttachmentTest.java
@@ -4,6 +4,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.attachment.Attachment;
+
 public class AttachmentTest {
 
     @Test
